🚚 Move component to the onboarding module
